+++ libraries/TCA9555/TCA9555.h
@@ -2,7 +2,7 @@
 //
 //    FILE: TCA9555.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.4
+// VERSION: 0.1.5
 // PURPOSE: Arduino library for I2C TCA9555 16 channel port expander
 //    DATE: 2021-06-09
 //     URL: https://github.com/RobTillaart/TCA9555
@@ -12,7 +12,7 @@
 #include "Wire.h"
 
 
-#define TCA9555_LIB_VERSION               (F("0.1.4"))
+#define TCA9555_LIB_VERSION               (F("0.1.5"))
 
 #define TCA9555_OK                        0x00
 #define TCA9555_PIN_ERROR                 0x81
@@ -21,6 +21,29 @@
 #define TCA9555_PORT_ERROR                0x84
 
 #define TCA9555_INVALID_READ              -100
+
+
+#if !defined(TCA9555_PIN_NAMES)
+#define TCA9555_PIN_NAMES
+
+  #define TCA_P00         0
+  #define TCA_P01         1
+  #define TCA_P02         2
+  #define TCA_P03         3
+  #define TCA_P04         4
+  #define TCA_P05         5
+  #define TCA_P06         6
+  #define TCA_P07         7
+  #define TCA_P10         8
+  #define TCA_P11         9
+  #define TCA_P12         10
+  #define TCA_P13         11
+  #define TCA_P14         12
+  #define TCA_P15         13
+  #define TCA_P16         14
+  #define TCA_P17         15
+
+#endif
 
 
 class TCA9555
@@ -71,6 +94,8 @@ public:
 
 
   int      lastError();
+  uint8_t  getType();
+
 
 protected:
   bool     writeRegister(uint8_t reg, uint8_t value);
@@ -79,6 +104,7 @@ protected:
   uint8_t  _address;
   TwoWire* _wire;
   uint8_t  _error;
+  uint8_t  _type;
 };